Citing fields

+++ classes/notes.php
@@ -6,10 +6,12 @@ class Note {
     private $author;
     private $note;
     private $table_id;
+    private $type;
+    private $saved;
 
 
     function __construct() {
-
+        $this->saved =false;
     }
 
     function set_data_from_row($row) {;
@@ -19,6 +21,14 @@ class Note {
         $this->set_note($row['note']);
         $this->set_table_id($row['table_id']);
 
+    }
+
+    function set_saved($saved) {
+        $this->saved = $saved;
+    }
+
+    function is_saved() {
+        return $this->saved;
     }
 
     function set_table_id($id) {
@@ -41,6 +51,14 @@ class Note {
         $this->note = $note;
     }
 
+    function set_type($type) {
+        $this->type = $type;
+    }
+
+    function get_type() {
+        return $this->type;
+    }
+
     function get_table_id() {
         return $this->table_id;
     }
@@ -61,12 +79,50 @@ class Note {
         return $this->id;
     }
 
-    function delete() {
-        global $notes;
-        $where = array('id' => $this->id);
-        $notes->delete(array('id' => $this->get_id()));
+    function save_note() {
 
-        return $notes->query();
+        if ($this->saved) {
+            return true;
+        }
+        global $notes;
+
+        $data = [
+            'date' => "STR_TO_DATE('".$this->get_date()."', '%Y-%m-%d %H:%i:%s')",
+            'table_id' => $this->get_table_id(),
+            'note' => $this->get_note(),
+            'author' => $this->get_author(),
+            'type' => $this->get_type()
+        ];
+
+            $notes->insert($data);  
+        
+
+        echo $notes->query;
+
+        if ($res = $notes->query()) {
+            $this->set_id($res);
+            $this->set_saved(true);
+            return true;
+        }
+
+        return false;
+
+    }
+
+    function delete() {
+        if ($this->saved) {
+            global $notes;
+            $where = array('id' => $this->id);
+            $notes->delete(array('id' => $this->get_id()));
+            if ($notes->query()) {
+                return true;
+            }
+
+            return false;
+        }
+        
+
+        return true;
     }
 }
